Moved TezosBalanceUpdater to own class. #95
