+++ Atomex.Client.Core/Services/Abstract/IChainBalanceUpdater.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading.Tasks;
+
+
+namespace Atomex.Services.Abstract
+{
+    public interface IChainBalanceUpdater
+    {
+        Task StartAsync();
+        Task StopAsync();
+    }
+}
