+++ Atomex.Client.Core/Services/BalanceUpdaters/TezosBalanceUpdater.cs
@@ -1,20 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading;
 using System.Threading.Tasks;
 using Atomex.Abstract;
 using Atomex.Services.Abstract;
-using Atomex.Services.BalanceUpdaters;
 using Atomex.TzktEvents;
-using Atomex.Wallet;
 using Atomex.Wallet.Abstract;
 using Serilog;
 
 
-namespace Atomex.Services
+namespace Atomex.Services.BalanceUpdaters
 {
-    public class BalanceUpdater : IBalanceUpdater
+    public class TezosBalanceUpdater : IChainBalanceUpdater
     {
         private readonly IAccount _account;
         private readonly ICurrenciesProvider _currenciesProvider;
@@ -22,98 +19,24 @@ namespace Atomex.Services
         private readonly ITzktEventsClient _tzktEvents;
         private readonly IHdWalletScanner _walletScanner;
 
-        private CancellationTokenSource _cts;
-        private bool _isRunning;
-
         private ISet<string> _tezosAddresses;
-        private readonly IList<IChainBalanceUpdater> _balanceUpdaters;
 
-        public BalanceUpdater(IAccount account, ICurrenciesProvider currenciesProvider, ILogger log)
+
+        public TezosBalanceUpdater(IAccount account, ICurrenciesProvider currenciesProvider, IHdWalletScanner walletScanner, ILogger log)
         {
             _account = account ?? throw new ArgumentNullException(nameof(account));
             _currenciesProvider = currenciesProvider;
             _log = log ?? throw new ArgumentNullException(nameof(log));
+            _walletScanner = walletScanner ?? throw new ArgumentNullException(nameof(walletScanner));
 
             _tzktEvents = new TzktEventsClient(_log);
-            _walletScanner = new HdWalletScanner(_account);
-
-            _balanceUpdaters = new List<IChainBalanceUpdater>();
-            InitChainBalanceUpdaters();
         }
 
-        public void Start()
-        {
-            if (_isRunning)
-            {
-                throw new InvalidOperationException("BalanceUpdater already running");
-            }
-
-            _isRunning = true;
-            _cts = new CancellationTokenSource();
-
-            Task.Run(async () =>
-            {
-                try
-                {
-                    foreach (var chainBalanceUpdater in _balanceUpdaters)
-                    {
-                        await chainBalanceUpdater.StartAsync().ConfigureAwait(false);
-                    }
-
-                    _log.Information("BalanceUpdater successfully started");
-                }
-                catch (OperationCanceledException)
-                {
-                    _log.Debug("BalanceUpdater canceled");
-                }
-                catch (Exception e)
-                {
-                    _log.Error(e, "Unconfirmed BalanceUpdater error");
-                }
-
-            }, _cts.Token);
-        }
-
-        public void Stop()
-        {
-            if (!_isRunning)
-            {
-                return;
-            }
-
-            Task.Run(async () =>
-            {
-                try
-                {
-                    foreach (var chainBalanceUpdater in _balanceUpdaters)
-                    {
-                        await chainBalanceUpdater.StopAsync().ConfigureAwait(false);
-                    }
-                }
-                catch (Exception e)
-                {
-                    _log.Error(e, "Error while stopping BalanceUpdater");
-                }
-            });
-
-            _cts.Cancel();
-            _isRunning = false;
-            _log.Information("BalanceUpdater stopped");
-        }
-
-        private void InitChainBalanceUpdaters()
-        {
-            _balanceUpdaters.Add(new TezosBalanceUpdater(_account, _currenciesProvider, _walletScanner, _log));
-        }
-
-
-        #region Tezos
-
-        private async Task StartTezosBalanceUpdater()
+        public async Task StartAsync()
         {
             if (_currenciesProvider == null)
             {
-                throw new InvalidOperationException("StartTezosBalanceUpdater was called before CurrenciesProvider initialization");
+                throw new InvalidOperationException("Start TezosBalanceUpdater was called before CurrenciesProvider initialization");
             }
 
             try
@@ -125,7 +48,7 @@ namespace Atomex.Services
 
                 await _tzktEvents.StartAsync(baseUri).ConfigureAwait(false);
                 _tezosAddresses = await GetAddressesAsync().ConfigureAwait(false);
-                
+
                 await _tzktEvents
                     .NotifyOnAccountsAsync(_tezosAddresses, TezosBalanceUpdateHandler)
                     .ConfigureAwait(false);
@@ -135,6 +58,12 @@ namespace Atomex.Services
                 _log.Error(e, "Error on starting TezosBalanceUpdater updater");
             }
         }
+
+        public async Task StopAsync()
+        {
+            await _tzktEvents.StopAsync().ConfigureAwait(false);
+        }
+
 
         private async Task<ISet<string>> GetAddressesAsync()
         {
@@ -148,8 +77,8 @@ namespace Atomex.Services
                 .ConfigureAwait(false);
 
             return addresses.Concat(new[] { freeAddress })
-                            .Select(wa => wa.Address)
-                            .ToHashSet();
+                .Select(wa => wa.Address)
+                .ToHashSet();
         }
 
         private async void TezosBalanceUpdateHandler(string address)
@@ -159,7 +88,7 @@ namespace Atomex.Services
                 await _walletScanner
                     .ScanAddressAsync(TezosConfig.Xtz, address)
                     .ConfigureAwait(false);
-                
+
                 var newAddresses = await GetAddressesAsync().ConfigureAwait(false);
                 newAddresses.ExceptWith(_tezosAddresses);
 
@@ -178,12 +107,5 @@ namespace Atomex.Services
                 _log.Error(e, "Error on handling Tezos balance update");
             }
         }
-
-        private async Task StopTezosBalanceUpdater()
-        {
-            await _tzktEvents.StopAsync().ConfigureAwait(false);
-        }
-
-        #endregion Tezos
     }
 }
