+++ Atomex.TzktEvents/TzktEventsClient.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using System.Xml.Serialization;
 using Microsoft.AspNetCore.SignalR.Client;
 using Serilog;
 using Atomex.TzktEvents.Services;
